+++ src/Attribute/FastlyApi.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace BEAR\FastlyModule;
+namespace BEAR\FastlyModule\Attribute;
 
 use Attribute;
 use Ray\Di\Di\Qualifier;
 
 #[Attribute(Attribute::TARGET_PARAMETER), Qualifier]
-final class ServiceId
+final class FastlyApi
 {
 }
