+++ src/FastlyPurgeModule.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace BEAR\FastlyModule;
 
+use BEAR\FastlyModule\Attribute\FastlyApi;
+use BEAR\FastlyModule\Attribute\ServiceId;
+use BEAR\FastlyModule\Attribute\SoftPurge;
 use BEAR\QueryRepository\PurgerInterface;
 use Fastly\Api\PurgeApi;
 use Fastly\Configuration;
